Put types into their modules. Refactor useStretchy
Now useStretchy is getting its config as an object.
So it has more clear api and cleaner code.
