+++ lib/src/hooks/useStretchy.ts
@@ -1,12 +1,30 @@
-import { UseStretchy } from '../types';
+import {
+  Animated,
+  NativeSyntheticEvent,
+  NativeScrollEvent,
+  LayoutChangeEvent,
+} from 'react-native';
+import { StretchyImage, StretchyOnScroll } from '../types';
 import { useImageHeightBasedOnRatio } from './useImageHeightBasedOnRatio';
 import { useOnScrollHandle } from './useOnScrollHandle';
 
-export const useStretchy: UseStretchy = (
+export type UseStretchy = (config: {
+  image?: StretchyImage;
+  preferredImageHeight?: number;
+  scrollListener?: StretchyOnScroll;
+}) => {
+  animation: Animated.Value;
+  heightBasedOnRatio: number;
+  onImageLoad: () => void;
+  onScroll(event: NativeSyntheticEvent<NativeScrollEvent>): void;
+  onImageWrapperLayout(event: LayoutChangeEvent): void;
+};
+
+export const useStretchy: UseStretchy = ({
   image,
   preferredImageHeight,
   scrollListener,
-) => {
+}) => {
   const { animation, onScroll, onImageWrapperLayout } = useOnScrollHandle(
     scrollListener,
   );
